requirements.txt
python-dotenv==0.10.1
Scrapy==1.8.0
Django==2.2.4
djangorestframework==3.10.2
django-health-check==3.11.1
django-rest-swagger==2.2.0
django-rest-auth==0.9.5
django-filter==2.2.0
postgres==2.2.2
psycopg2-binary==2.8.3
ipython==7.9.0
raven==6.10.0
swagger-spec-validator==2.4.3
Faker==2.0.4
coverage==3.6
scrapy-djangoitem==1.1.1
